+++ src/harv.c
@@ -16,22 +16,35 @@ struct h_data {
 	short int min[9];
 	short int max[9];
 	short int unharvest_prob;
+	short int reharvest_turns;
+	short int reharvest_prob;
 } H_TABLE[] = {
-	{ 0, 1, 4, { 1, 2, 3, 4, 0, 0, 0, 0, 0 }, { 25, 25, 25, 25, 0, 0, 0, 0, 0 },
-	           {15, 1, 1, 1, 0, 0, 0, 0, 0}, {100, 3, 1, 1, 0, 0, 0, 0, 0}, 25 },
+	{ 0, 1, 6, { 1, 2, 3, 4, 0, 0, 0, 0, 0 }, { 25, 25, 25, 25, 0, 0, 0, 0, 0 },
+	           {15, 1, 1, 1, 0, 0, 0, 0, 0}, {100, 3, 1, 1, 0, 0, 0, 0, 0}, 25, 10, 5 },
 	{ 0, 2, 4, { 5, 6, 7, 8, 0, 0, 0, 0, 0 }, { 11, 11, 11, 67, 0, 0, 0, 0, 0 },
-		   {1, 1, 1, 1, 0, 0, 0, 0, 0}, {5, 1, 1, 1, 0, 0, 0, 0, 0}, 50 },
+		   {1, 1, 1, 1, 0, 0, 0, 0, 0}, {5, 1, 1, 1, 0, 0, 0, 0, 0}, 50, 20, 5 },
+	{ 0, 3, 2, { 9, 10, 0, 0, 0, 0, 0, 0, 0 }, { 50, 50, 0, 0, 0, 0, 0, 0, 0 },
+		   { 1, 1, 0, 0, 0, 0, 0, 0, 0 }, { 1, 1, 0, 0, 0, 0, 0, 0, 0 }, 50, 20, 5 },
 	{ -1, 0, 0, { 0, 0, 0, 0, 0, 0, 0, 0, 0 }, { 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-		   { 0, 0, 0, 0, 0, 0, 0, 0, 0 }, { 0, 0, 0, 0, 0, 0, 0, 0, 0 }, 0 }
+		   { 0, 0, 0, 0, 0, 0, 0, 0, 0 }, { 0, 0, 0, 0, 0, 0, 0, 0, 0 }, 0, 0, 0 }
 };
+
+struct d_data {
+	struct d_data *next;
+	int x;
+	int y;
+	short int turn;
+	short int prob;
+} *D_TABLE;
 
 /* Function prototypes */
 static short int	get_item(short int biome, short int tile);
 static unsigned char	get_quantity(short int biome, short int tile, short int item_id);
 static void		make_unharvestable(struct worldmap *map, int x, int y, short int biome, short int tile);
-static SDL_bool		stash_item(struct worldmap *map, struct player *cur_player, int x, int y, short int item_id, unsigned char quantity);
+static SDL_bool		stash_item(struct player *cur_player, short int item_id, unsigned char quantity);
 static void		drop_item(struct worldmap *map, int x, int y, short int item_id, unsigned char quantity);
 static SDL_bool		harvest_loop(struct game *cur_game, struct worldmap *map, struct player *cur_player);
+static void		add_dtable(struct worldmap *map, int x, int y);
 
 void
 get_harvest_input(struct game *cur_game, struct worldmap *map, struct player *cur_player)
@@ -96,11 +109,20 @@ harvest_item(struct game *cur_game, struct worldmap *map, struct player *cur_pla
 		*(*(map->quantity+y)+x) = 0;
 		*(*(map->loot+y)+x) = 0;
 		/* Try to stash the item and drop the remainder */
-		if (stash_item(map, cur_player, x, y, item, quantity) == SDL_FALSE) {
+		if (stash_item(cur_player, item, quantity) == SDL_FALSE) {
 			drop_item(map, x, y, item, quantity);
 		}
 		return SDL_TRUE;
-	}
+	} else if (*(*(map->ground+y)+x) != 0) {
+		/* Is there a ground on the map? */
+		item = *(*(map->ground+y)+x);
+		/* Try to stash the ground and take off map if stash is successful */
+		if (stash_item(cur_player, item, 1) == SDL_TRUE) {
+			/* Take ground off map */
+			*(*(map->ground+y)+x) = 0;
+		}
+		return SDL_TRUE;
+	}	
 	
 	/* No item on map, try to harvest */
 	biome = *(*(map->biome+y)+x);
@@ -112,7 +134,8 @@ harvest_item(struct game *cur_game, struct worldmap *map, struct player *cur_pla
 			finished = harvest_loop(cur_game, map, cur_player);
 			if (finished == SDL_TRUE) {
 				quantity = get_quantity(biome, tile, item);
-				if (stash_item(map, cur_player, x, y, item, quantity) == SDL_FALSE) {
+				if (stash_item(cur_player, item, quantity) == SDL_FALSE) {
+					drop_item(map, x, y, item, quantity);
 					finished = SDL_FALSE;
 				}
 				make_unharvestable(map, x, y, biome, tile);
@@ -192,11 +215,12 @@ make_unharvestable(struct worldmap *map, int x, int y, short int biome, short in
 	rando = rand_num(0, 99);
 	if (rando < H_TABLE[i].unharvest_prob) {
 		*(*(map->harvestable+y)+x) = SDL_FALSE;
+		add_dtable(map, x, y);
 	}
 }
 
 static SDL_bool
-stash_item(struct worldmap *map, struct player *cur_player, int x, int y, short int item_id, unsigned char quantity)
+stash_item(struct player *cur_player, short int item_id, unsigned char quantity)
 {
 	int i;
 	
@@ -225,7 +249,6 @@ stash_item(struct worldmap *map, struct player *cur_player, int x, int y, short 
 		}
 	}
 	/* Could not put item in inventory */
-	drop_item(map, x, y, item_id, quantity);
 	return SDL_FALSE;
 }
 
@@ -261,7 +284,21 @@ harvest_loop(struct game *cur_game, struct worldmap *map, struct player *cur_pla
 	while (finished == SDL_FALSE) {
 		/* Check for input */
 		SDL_PollEvent(&event);
-		if (event.type == SDL_KEYDOWN || event.type == SDL_MOUSEBUTTONDOWN) {
+		if (event.type == SDL_KEYDOWN) {
+			if (event.key.keysym.sym == SDLK_i) {
+				/* toggle inventory and re-render screen */
+				toggle_inv(cur_game);
+				SDL_DestroyTexture(texture);
+				texture = SDL_CreateTexture(cur_game->screen.renderer, SDL_PIXELFORMAT_RGBA8888,
+							    SDL_TEXTUREACCESS_TARGET, WIN_W, WIN_H);
+				SDL_SetRenderTarget(cur_game->screen.renderer, texture);
+				draw_game(cur_game, map, cur_player);
+				draw_rect(cur_game, GAME_X, GAME_Y, GAME_W, GAME_H, SDL_FALSE, white, SDL_FALSE, NULL);
+				SDL_SetRenderTarget(cur_game->screen.renderer, NULL);
+			} else {
+				return SDL_FALSE;
+			}
+		} else if (event.type == SDL_MOUSEBUTTONDOWN) {
 			return SDL_FALSE;
 		}
 		/* Output screen */
@@ -284,4 +321,91 @@ harvest_loop(struct game *cur_game, struct worldmap *map, struct player *cur_pla
 	}
 	SDL_DestroyTexture(texture);
 	return SDL_TRUE;
+}
+
+void
+setup_dtable(void)
+{
+	/* Make an initial entry in D_TABLE */
+	D_TABLE = malloc(sizeof(*D_TABLE)*1);
+	D_TABLE->next = NULL;
+}
+
+static void
+add_dtable(struct worldmap *map, int x, int y)
+{
+	short int biome;
+	short int tile;
+	short int turn;
+	short int prob;
+	int i;
+	struct d_data *current;
+	struct d_data *tmp;
+
+	/* Look up turn and prob */
+	biome = *(*(map->biome+y)+x);
+	tile = *(*(map->tile+y)+x);
+	for (i = 0; H_TABLE[i].biome != -1; i++) {
+		if (H_TABLE[i].biome == biome && H_TABLE[i].tile == tile) {
+			turn = H_TABLE[i].reharvest_turns;
+			prob = H_TABLE[i].reharvest_prob;
+			break;
+		}
+	}
+	if (H_TABLE[i].biome == -1) return;
+					
+	/* Set values in current d_data */
+	current = malloc(sizeof(*current)*1);
+	current->x = x;
+	current->y = y;
+	current->turn = turn;
+	current->prob = prob;
+	current->next = NULL;
+	
+	/* Add to linked list */
+	tmp = D_TABLE;
+	while (SDL_TRUE) {
+		if (tmp->next == NULL) {
+			tmp->next = current;
+			return;
+		} else {
+			tmp = tmp->next;
+		}
+	}
+}
+
+void
+check_depleted(struct worldmap *map)
+{
+	int rando;
+	struct d_data *tmp;
+	struct d_data *last;
+	
+	/* Run through the D_TABLE and check if tile needs to regenerate */
+	tmp = D_TABLE;
+	while (tmp->next != NULL) {
+		last = tmp;
+		tmp = tmp->next;
+		if (tmp->turn > 0) {
+			/* If tile is not ready to regenerate, adjust turns */
+			tmp->turn--;
+		} else {
+			/* Tile is ready to regenerate, check probability */
+			rando = rand_num(0, 99);
+			if (rando < tmp->prob) {
+				/* Regenerate the tile */
+				*(*(map->harvestable+tmp->y)+tmp->x) = SDL_TRUE;
+				/* Remove from linked list */
+				last->next = tmp->next;
+				free(tmp);
+				tmp = last;
+			}
+		}
+	}
+}
+
+void
+kill_dtable(void)
+{
+
 }