+++ src/maps.c
@@ -130,17 +130,20 @@ create_map(struct worldmap *map, int row_size, int col_size)
 	map->loot = malloc(sizeof(*(map->loot))*row_size);
 	map->quantity = malloc(sizeof(*(map->quantity))*row_size);
 	map->harvestable = malloc(sizeof(*(map->harvestable))*row_size);
+	map->ground = malloc(sizeof(*(map->ground))*row_size);
 	for (rows = 0; rows < row_size; rows++) {
 		*(map->tile+rows) = malloc(sizeof(**map->tile)*col_size);
 		*(map->biome+rows) = malloc(sizeof(**map->biome)*col_size);
 		*(map->loot+rows) = malloc(sizeof(**map->loot)*col_size);
 		*(map->quantity+rows) = malloc(sizeof(**map->quantity)*col_size);
+		*(map->ground+rows) = malloc(sizeof(**map->ground)*col_size);
 		*(map->harvestable+rows) = malloc(sizeof(**map->harvestable)*col_size);
 		for (cols = 0; cols < col_size; cols++) {
 			*(*(map->tile+rows)+cols) = 0;
 			*(*(map->biome+rows)+cols) = 0;
 			*(*(map->loot+rows)+cols) = 0;
 			*(*(map->quantity+rows)+cols) = 0;
+			*(*(map->ground+rows)+cols) = 0;
 			*(*(map->harvestable+rows)+cols) = SDL_TRUE;
 		}
 	}
@@ -157,12 +160,14 @@ free_map(struct worldmap *map)
 		free(*(map->biome+rows));
 		free(*(map->loot+rows));
 		free(*(map->quantity+rows));
+		free(*(map->ground+rows));
 		free(*(map->harvestable+rows));
 	}
 	free(map->tile);
 	free(map->biome);
 	free(map->loot);
 	free(map->quantity);
+	free(map->ground);
 	free(map->harvestable);
 }